Add files via upload
1)Brief explanation

Ferromagnetic materials such as Iron respond to a magnetic field by aligning their microscopic magnetic moments in the same direction, such that a significantly large net magnetisation is measured. This magnetisation eventually reaches a saturation point and does not disappear when the applied field is removed. The applied field can then be used in the opposite direction to gradually reduce and flip the sample’s magnetisation in the opposite direction. The sample’s magnetisation can therefore undergo a full “loop” if the magnetic field is applied correctly. This is what is known as the hysteresis loop.

This project compared the model and the data for the magnetic response of crystalline nanoparticle film alloys composed of Iron (Fe) and Aluminium (Al) under the application of an external magnetic field. Four samples were tested against the model, each prepared by gas-phase deposition techniques and containing different concentrations of Iron to Aluminium. The model seems to fit Samples 1 and 2 the best, while it appears to fit Sample 4 the worst.

2)Data

The data for each sample represents the sample's magnetisation as a function of the strength and direction of the applied magnetic field. Zone 1 refers to all the magnetisation values observed when applying the magnetic field in the positive direction, while zone 2 refers to the magnetisation observed when applying the magnetic field in the opposite direction